Identity Server API , Branch API, Account API Account Access API review OK

+++ Src/Security/Identity.API/Services/EntitlementService.cs
@@ -3,6 +3,7 @@ using Identity.API.Applications.Dtos;
 using Identity.API.Applications.Models.Entities;
 using Identity.API.Applications.Models.Responses;
 using Identity.API.Services.Interfaces;
+using Identity.API.Utils;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.VisualBasic;
 using System.Data;
@@ -41,13 +42,19 @@ namespace Identity.API.Services
             IEnumerable<string> roles = new List<string>() { role.Name };
             await this._userManager.AddToRolesAsync(user, roles);
 
-            return  GetEntitlementResponseFromModelAsync(await _roleManager.FindByNameAsync(request.Role_name));
+            return  GetEntitlementResponseFromModelAsync(await _roleManager.FindByNameAsync(role.Name));
         }
 
         public async Task<bool> DeleteEntitlementAsync(string entitlement_id)
         {
-            var role = await _roleManager.FindByIdAsync(entitlement_id);
+            var role = _context.Entitlements.Where( e => e.Id == entitlement_id && !SystemRoles.Roles.Contains(e.Name)).First();
+            if (role == null)
+            {
+                return false;
+            }
+       
             IdentityResult result = await _roleManager.DeleteAsync(role);
+            //_userManager.rem
             if (result.Succeeded)
             {
                 return true;
@@ -58,7 +65,7 @@ namespace Identity.API.Services
         public Entitlements GetAllEntitlements()
         {
             IList<EntitlementResponse> list = new List<EntitlementResponse>();
-            var roles = _context.Entitlements.ToList();
+            var roles = _context.Entitlements.Where(r => !SystemRoles.Roles.Contains(r.Name)).ToList();
             foreach (var item in roles)
             {
                 list.Add(GetEntitlementResponseFromModelAsync(item));
@@ -66,8 +73,32 @@ namespace Identity.API.Services
             return new Entitlements() { list = list };  
         }
 
+        public async Task<Entitlements> GetEntitlementForUserAsync(string userId)
+        {
+            IList<EntitlementResponse> list = new List<EntitlementResponse>();
+            var user = await _userManager.FindByIdAsync(userId);
+            if (user != null)
+            {
+                var roles = await _userManager.GetRolesAsync(user);
+                foreach (var item in roles.Where(r => !SystemRoles.Roles.Contains(r)))
+                {
+                    var role = _context.Entitlements.Where( r => r.Name == item).FirstOrDefault();
+                    list.Add(GetEntitlementResponseFromModelAsync(role));
+                }
+                return new Entitlements() { list = list };
+            }
+            else
+            {
+                return new Entitlements();
+            }
+        }
+
         private EntitlementResponse GetEntitlementResponseFromModelAsync(Entitlement itemRole)
         {
+            if (itemRole == null)
+            {
+                return new EntitlementResponse();
+            }
             EntitlementResponse entitlement = new EntitlementResponse()
             {
                 Bank_id = itemRole.Bank_id,
